Convert the quake dialect ops to support both memory and register forms. (#124)
* Convert the quake dialect ops to support both memory and register forms.
This unifies the quantum gate definitions into a single Op and
eliminates the overhead of supporting parallel dialects for the same
operations. See issue #98 for more.

A large portion of these changes involves converting the tests to
support a new syntax in which types are mandatory on quantum operations
and more uniformly follow the MLIR default functional notation.

Modify some specific points in the existing code to use the new op
builder routines. Also begin the work of phasing out the Common directory,
which will become obsolete after this refactoring is complete.

* Update convert_ctrl documentation.

Drop the "q" prefix from the control type. In forthcoming work, the
extra "q" prefixes on types in the dialect will be removed.

* Repair the python tests. More changes to check lines.

* Fix bugs in decompositions now that result values are variadic.