Revision of the repo check workflows (#131)
This PR refactors the spell and link checks, adds a workflow to check the formatting of C++ and Python files, and a check that all files have a copyright header. It creates a new workflow with various repo checks that replaces the old spell and link checker, and adds header and formatting checks.

Spell checking for C++: checks all comments in header files as well as cpp files in the examples. Content between backticks (inline or as block) is treated as code and ignored. Strings are unchecked.
Spell checking in Python: checks all comments, doc comments, and strings in python files (source code and examples). Content between backticks (inline or as block) is treated as code and ignored.
Markdown files: Code is ignored, link titles are checked.
Rst files: Processed like markdown files, with some additional rules to filter lines starting with two dots.

Not currently covered: Docs strings of the code generated by PyBind, since C++ strings are not spell checked.