[opt] Add memtoreg and regtomem passes. (#233)
* [opt] Add memtoreg and regtomem passes.

The memtoreg pass is a generalized pass to iteratively convert Quake and
CC dialects from a memory reference semantics to a register value
semantics. Quake has some different properties than the classical case,
but through a preparatory transformation to QLS format, both the quantum
and the classical code can ultimately use the same framework to thread
wire values as well as classical scalars.

There is also a regtomem pass that only works on Quake. This re-wraps
the wire values (the register value semantics) as references, if
possible, and returns the Quake IR to the qubit (memory) reference
semantics.

Add tests.

* Review comments.