[pass] Add a loop normalization pass. (#313)
* [pass] Add a loop normalization pass.

Implement enhancement #288.

This pass converts monotonic loops with constant step into invariant
loops or counted loops depending on whether the total number of
iterations is constant. This pass has some limitations which can be
addressed in future work. For now it will convert some interesting loops
such as countdown loops, loops with steps != 1, etc.

Normalizing loops is an important transformation in that it will allow
the loop unrolling pass to find more counted loops to unroll.

* Fix documentation.

* Remove unsigned division. When computing the number of iterations in a
countdown loop, we want signed division. Specifically, the calculation
of (upper - lower + step), where upper < lower and step is negative,
will produce a negative value. Signed division yields either a positive
value (when we have iterations to execute) or a negative value when the
preconditions are not true.

* Add comment to explain why checks are required in the match and rewrite.

Add more test cases to check that the math to compute the number of
iterations, including negative cases, does not regress.

* Code cleanup on the loop unroll pass. Eliminate some redundant checks
which are no longer needed with the use of the loop analysis framework.

* Make the option name "signal-failure" longer to fully specify what it does and hopefully eliminate misinterpretation.