Workflow updates to enable multi-platform docker images and cuda quantum packages with additional components (#322)
Completes the work for #41.

This PR:

Migrates all docker builds that create public images to use the build-and-push action and switches to use the docker container driver everywhere. This enables multi-platform builds in the future, but requires that we push images to a local registry if we don't push them to a public one. See also the docs.
Pushes the images to a local registry during CI. However, we currently don't share the local registry between jobs. Multi-platform builds are hence currently possible during deployment, but not during CI. This feature may facilitate multi-platform builds also during CI without dumping the registry.
Creates both a release and a debug build of CUDA Quantum for as part of CI. The debug build is included in the cuda-quantum-dev package that is published on ghcr. The release build is used to build the cuda-quantum docker image.
Makes the necessary changes so that GitLab artifacts can be pushed to a release draft.
I took the liberty of directly adding a workflow to main to clean up deployments to default (see the comment here for context), and to update the GitLab workflow on bot/data to confirm everything works as intended.
Adds a GitHub workflow that takes the GitLab (and other) artifacts in a release and creates a cuda-quantum image that contains these as well.
Added a manual trigger to have the workflow also create a complete GitHub release.