Tag changes for cuda-quantum image and signing (#561)
The tags for the cuda-quantum image change as follows:
- the tag `latest-hpc` is replaced by the `latest` tag
- the tag `latest` is replaced by the `latest-base` tag

If the NGC_CREDENTIALS are defined as a secret in the ghcr-deployment environment, the pipelines also push the cuda-quantum `latest` and `latest-base` images to NGC, as well as the version from tags and (future) release branches. 

Container signing is done using cosign, see also https://github.com/sigstore/cosign. More information in the description of this PR (#561).