Only create tar cache for dev dependencies when they are modified (#599)
This PR adds a step to the dev_environment.yml workflow that checks if the latest image on the GHCR registry match the layers needed for the build. If they do, it omits pushing them into the GH cache to reduce cache usage and instead uses the image from the registry in subsequent steps.