Fix #627: Code gen for IQM Json failure. (#632)
* Fix #627: Code gen for IQM Json failure.

These changes allow all the code gen backends to use a pass pipeline to
fix up the code before the actual file is written. In the case of the
bug, there was still Quake IR sitting in the Module that needed to be
transformed before it could be emitted correctly.

* Oops. Change so that inner lambdas don't disappear with the outer lambda
closure.