Modify Quake::ApplyOp to allow an indirect form of call in order to (#663)
* Modify Quake::ApplyOp to allow an indirect form of call in order to
allow application to callables, such as lambda expressions.

Teach the bridge to lower lambda expressions to indirect apply calls
when the lambda is not explicitly flagged with a `__qpu__` attribute but
is an available lambda expression. (There is no reason not to allow this
since the lambda expression is lowered by the bridge and readily
available.)

Augment the cc.instantiate_callable operation to allow for wrapping
plain old functions. Plain old functions are callable, but they do not
have captures, do not need to create a closure, and do not need a thunk
to unpack a closure. In short, they are a less complicated form.
Supporting them now allows two flavors of CallableType. The original
form, as constructed by lambda lifting, generates a pair of pointers
consisting of a thunk function pointer and a closure pointer. The new
degenerate form consists of a pair of pointers consisting of the plain
old function pointer and a null pointer.

Add support for cc.call_callable to the code gen to QIR. In the event
that an SSA value of CallableType is an indeterminant value, code gen
will generate the code to test between the thunk form and the plain old
function form, as described above, and generate the correct call to
either the thunk function or directly to the regular function itself.
Variants of this can be found in the test.

Update documentation.

Add tests.

* Adding missing CHECK lines.