Move unwind lowering to early in the pipeline and remodel memtoreg pass (#764)
* Change the driver to lower the unwind* pseudo terminators early.

The unwind pseudo terminators are non-local control flow operations
(gotos) that help the bridge lower C++ to the CC dialect. Since they
imply global (intra-function) transfers of control (and as such have
special semantics and are not actual terminators), they are harder to
reason about. This change moves the pass that transforms the unwind ops
to CFG control flow to the very front of the nvq++ pipeline to erase
them as early as possible.

* Renovate the memtoreg pass.

These changes fix some bugs in the pass but are also, largely, a big
refactoring of the pass. The code now uses a data-flow class to track
the various kinds and modes of data-flow information used and tracked to
convert an MLIR-like memory-semantics SSA form to a register-semantics
SSA form. This includes more documentation on the various data
structures as well as assertions to check they are not misused.

* Add tests.

* Update lib/Optimizer/Transforms/MemToReg.cpp

Co-authored-by: Ben Howe <141149032+bmhowe23@users.noreply.github.com>

---------

Co-authored-by: Ben Howe <141149032+bmhowe23@users.noreply.github.com>