Extend the cc-loop-unroll pass. (#768)
Add an option to the unrolling pass and unrolling pass pipeline to allow
loops with `break` statements to be fully unrolled, so long as the loop
is otherwise, structurally speaking, like a regular counted loop. Since
the total number of executions is a variable, the option allows one to
selectively enable/disable the unrolling of these variable iteration,
constant upper bound loops.

Fix bug with default for testing for an invariant loop.

Guard against unrolling a loop with zero iterations.