Implement std::reverse on std::vector (#836)
* An implementation for std::reverse on a std::vector.

This is a preliminary implementation of std::reverse(). The function
expects two iterators, which are pointers in degenerated form. This code
performs a pair-wise swap of the elements of the vector in memory.

This implementation won't optimize well at the moment because of the
pointer arithmetic stemming from the iterators. Specifically, the
optimizer won't be able to unroll this loop. We should consider
implementing something that uses the vector's size and/or offset value,
when they are constant.

Add test.

Add CHECK-DAGs to allow constants to appear in different orders when
running the test.

* Update test for new canonicalization patterns.