Simplify lambda expressions. (#950)
* Simplify lambda expressions.

These changes simplify some of the lambda expressions in the cudaq
headers. It eliminates some of the more complicated capture expressions
and uses simple reference capture. It also removes the `mutable`
modifier which does not appear to be necessary in these cases.

* Revert some of the changes. It looks like there is a bug in GCC 12 that
is resulting in a test failure in qpp_AsyncTester.checkSampleAsync when
building Debug mode. I've tried multiple ways of writing the lambdas in
question, which are surprisingly fairly simple, without success.