Fix #1348 - synthesis of vectors with small integers. (#1395)
The constant array op is using at ArrayAttr, which does not naturally
support integral sizes less than i32. This patch upcasts the short
values to i32 to get around the issue. Another approach would be to
have the constant array op take a DenseArrayAttr, which does support
the shorter integral types. Might need a custom parser/pretty printer
if we go that route?