[c++17] changes to cudaq headers for pure c++17 environment (#1468)
* Disable the C++20 template form of adding controls to gates.

We were allowing calls such as
  x<cudaq::ctrl>(q);
in C++17 compilation mode even though that syntax requires C++20. This
PR prevents using the C++20 syntax when compiling as C++17. C++17 code
will have to rewrite the above as
  cx(q);
More generally, the cudaq::control and cudaq::adjoint functions can be
used.

Change the definition of my to be C++17.

Refactor the definition of swap to be C++17. Adds a "cswap" for the
control variant.

Add test.

Add handling of cr[xyz1] to the bridge.

Port the conditional_sample.cpp test to C++17.

* Mark tests that use C++20 syntax as REQUIRES c++20.

Port negation.cpp test to have a c++17 variant.

* Changes per review comments.