Blochsphere Visualization (#1763)
Fixes #1639 by adding a visualization for Bloch sphere for single-qubit states.
Significant changes:

Added QuTiP as an optional dependency in pyproject.toml
Added bloch_visualize module in python/cudaq/visualization with code to handle Bloch sphere generation and management. Docstrings are contained within the code itself. Tested with Jupyter notebooks.
Modified Python initialization init.py to import QuTiP, and if successful, load the visualization functionality.
REST server cannot be built on Singularity containers because CMake is never able to find ZLib/Minizip, especially if the program must search manually. Corrected the corresponding CmakeLists.txt so it searches in the correct places.
This is a submission for Unitary Hack 2024.