[core] Fix semantic problems with cc.compute_ptr (#1853)
* [core] Fix semantic problems with cc.compute_ptr

The CC dialect has a strong semantics on sequence types. Unlike C (and
LLVM), a pointer to a scalar type is not conflated with a pointer to an
array of scalars. The types are explicitly distinct.

As a consequence, the following has no semantics:

```mlir
  %1 = ... : !cc.ptr<i32>
  %2 = cc.compute_ptr %1[5] : (!cc.ptr<i32>) -> !cc.ptr<i32>
```

This construct violates the semantics of cc.compute_ptr and will now
fail to pass the verifier. The correct way to do this is:

```mlir
  %1 = ... : !cc.ptr<!cc.array<i32 x ?>>
  %2 = cc.compute_ptr %1[5] : (!cc.ptr<!cc.array<i32 x ?>>) -> !cc.ptr<i32>
```

These patches add the checks and fix all the places where the semantics
were being violated.

As a consequence, it will be easier to reason about cc.compute_ptr, to
debug issues, and to write code that uses cc.compute_ptr.

Reimplement the quake add metadata pass so that it isn't doing ad hoc
pattern matching and is prone to failure when small changes to the IR
happen. (Is there an issue for this?)

Add more canonicalizations for cc.compute_ptr.

Fix the python tests with FileCheck.

* Update new test.