[core] Add -lift-to-array-value pass (#1869)
* [core] Add -lift-to-array-value pass

This pass will promote an array constructed inline on the stack to a
cc.const_array op.

TODO: when the array is used by quake.init_state, the verifier will
fail. Nothing in the code base is built to accept an array *value* as an
input to quake.init_state's second argument.

* Add dominance checks for correctness.

* Fix bug with value vs. pointer and init_state.
Convert array constants to global variables in any case that the
constant array can't be decomposed into straight up constant
propagation.
Add test.