[core] Add the concept of wire sets. (#1864)
* [core] Add the concept of wire sets.

A wire set is a new abstraction in Quake. A wire set is a finite set of
wires (a.k.a. qubits). Each wire has a unique identity. All wires remain
linear types.

The usage model is to borrow a wire from a set, use it, and return it to
the set.

A wire set can also model other physical properties of qubits such as
topology via a sparse adjecency map (for copious generality, including
unidirectional links).

* Work around warning.

* Implement the verification of BorrowWire.

* Add documentation and comments.

"Pin" the value semantics instructions to prevent code motion from
moving operations when there is a mix of value and reference semantics.