[runtime] Fix bug with function declarations being dropped on the floor (#1919)
This small change copies function declarations from the "old" module to
the new module to fix correctness issues. This will never be a problem
if the old module is well formed. Dropping them on the floor is a bug
because there may be calls to these functions in the selected kernel,
and those calls will fail to verify after the declarations are dropped.