Fix #17: don't generate loops in the kernel_builder for measurements (#1910)
* Fix #17: don't generate loops in the kernel_builder for measurements

This simplifies kernel_builder and leverages that Quake already knows
what to do with veq types.

Fix reset while we are here. There is no need to test the types of these
arguments. The verifier already does that.

Fix std::string arguments to not make copies of the string at each call
site.

* Back out change to reset. Appears to choke in codegen. Will be a
separate PR.

* Add expand measurements pass to IQM pipeline.

* Add unrolling as well.