NVQC Optimizations for VQE (C++ and Python) (#1901)
This set of changes optimizes VQE performance when running simulations on NVQC.

The changes work for both C++ and Python applications.

Details: allow cudaq::vqe() to be invoked with variadic arguments similar to how we invoke cudaq::sample() and cudaq::observe(). More specifically, that means that we allow the user to pass the concrete, non-variational arguments (the arguments not subject to VQE optimization) directly into the cudaq::vqe() call. Users must switch to this calling convention to achieve notable speedups.