[runtime] Implements argument conversions as a runtime library. (#2040)
* [runtime] Implements argument conversions as a runtime library.

These changes add a new `ArgumentConverter` class which will be
used to translate argument values to MLIR code. The code is
meant to be used as a library by the runtime. Argument data will
arrive as raw pointers, so this uses auxiliary type information
to decode the arguments and translate them into constants as
MLIR code that are supported and used on the device side.

* add clang-format off/on to avoid clang-format breaking CHECK lines.

* Fix warning about truncation by making it explicit.

* Update runtime/common/ArgumentConversion.h

Co-authored-by: Thien Nguyen <58006629+1tnguyen@users.noreply.github.com>

* Review comments.

---------

Co-authored-by: Thien Nguyen <58006629+1tnguyen@users.noreply.github.com>